Replace GoalsTableViewController in storyboard
Replaced the Goals table in code with a generic view controller that accepts a table, but it would not allow its type to be changed to something other than inheriting from UITableView. So rebuilt the view from scratch. Also added a toolbar at bottom, to get Apple's built in blur effect. A user feedback label sits above that.